requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
celery==5.6.2
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
cssselect2==0.8.0
defusedxml==0.7.1
Django==6.0.1
django-celery-results==2.6.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-global-permissions==0.2.6
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-nested-routers==0.95.0
exceptiongroup==1.2.2
fonttools==4.57.0
google-api-core==2.25.0rc0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
holidays==0.66
html5lib==1.1
idna==3.10
iniconfig==2.1.0
lxml==5.4.0
oauthlib==3.2.0
oscrypto==1.3.0
packaging==25.0
phonenumbers==8.13.54
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pypdf==5.4.0
PyPDF2==3.0.1
pyphen==0.17.2
pytest==8.3.5
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.14.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.1
redis==7.1.0
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.12.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1